+++ lib/msf/core/exploit/mixins.rb
@@ -98,4 +98,7 @@ require 'msf/core/exploit/winrm'
 # WebApp
 require 'msf/core/exploit/web'
 
+# Firefox addons
+require 'msf/core/exploit/remote/firefox_addon_generator'
+
 require 'msf/core/exploit/remote/browser_exploit_server'
